requirements.txt
streamlit==1.31.0
python-dotenv==1.0.1
pypdf==4.0.1
langchain==0.1.11
langchain-core==0.1.30
langchain-mistralai==0.0.4
faiss-cpu==1.7.4
sentence-transformers==2.3.1